replace httparty/httmultiparty with rest-client
httparty has some serious issues when it comes to custom accept
headers. the request body is not correctly sent when using a versioned
accept header like "application/vnd.crowbar.v2.0+json"

we have been using httmultiparty, which is a thin layer above httparty
mainly to enable multi part file uploads.
this is supported by default in rest-client.

the only thing that was missing in rest-client was digest
authentication,
which is now added through a patch in our package that contains
https://github.com/rest-client/rest-client/pull/520

futhermore rest-client is actively maintained