make the check_id server agnostic
to not be dependent on the server response when it comes to a check
id, try to extract the check_id from the server error response.

the error response is wrapped into a check_id as well and it is safe
to extract it from there in case that it exists.

on top of that it should be insured that every error response is
wrapped into the same check_id as the top level check_id to provide
additional security.